sync

+++ sysmerge/sysmerge.sh
@@ -1,6 +1,6 @@
 #!/bin/ksh -
 #
-# $OpenBSD: sysmerge.sh,v 1.110 2014/01/22 10:40:07 ajacoutot Exp $
+# $OpenBSD: sysmerge.sh,v 1.118 2014/02/03 09:11:14 ajacoutot Exp $
 #
 # Copyright (c) 2008-2014 Antoine Jacoutot <ajacoutot@openbsd.org>
 # Copyright (c) 1998-2003 Douglas Barton <DougB@FreeBSD.org>
@@ -20,18 +20,28 @@
 
 umask 0022
 
-unset AUTO_INSTALLED_FILES BATCHMODE DIFFMODE ETCSUM NEED_NEWALIASES
-unset NEWGRP NEWUSR NEED_REBOOT NOSIGCHECK SRCDIR SRCSUM TGZ XETCSUM
-unset XTGZ
+unset AUTO_INSTALLED_FILES BATCHMODE DIFFMODE EDIT ETCSUM NEED_NEWALIASES
+unset NEWGRP NEWUSR NEED_REBOOT NOSIGCHECK RELINT SRCDIR SRCSUM TGZ
+unset XETCSUM XTGZ
 
+# forced variables
 WRKDIR=$(mktemp -d -p ${TMPDIR:=/var/tmp} sysmerge.XXXXXXXXXX) || exit 1
 SWIDTH=$(stty size | awk '{w=$2} END {if (w==0) {w=80} print w}')
+RELINT=$(uname -r | tr -d '.')
+if [ -z "${VISUAL}" ]; then
+	EDIT="${EDITOR:=/usr/bin/vi}"
+else
+	EDIT="${VISUAL}"
+fi
+
+# sysmerge specific variables (overridable)
 MERGE_CMD="${MERGE_CMD:=sdiff -as -w ${SWIDTH} -o}"
-FETCH_CMD="${FETCH_CMD:=/usr/bin/ftp -V -m -k ${FTP_KEEPALIVE:-0}}"
 REPORT="${REPORT:=${WRKDIR}/sysmerge.log}"
 DBDIR="${DBDIR:=/var/db/sysmerge}"
 
+# system-wide variables (overridable)
 PAGER="${PAGER:=/usr/bin/more}"
+SIGHASH=SHA256
 
 # clean leftovers created by make in src
 clean_src() {
@@ -58,11 +68,7 @@ usage() {
 }
 
 warn() {
-	echo "\t*** WARNING: $@"
-}
-
-error() {
-	echo "\t*** ERROR: $@"
+	echo "**** WARNING: $@"
 }
 
 report() {
@@ -71,10 +77,10 @@ report() {
 
 # remove newly created work directory and exit with status 1
 error_rm_wrkdir() {
-	(($#)) && error "$@"
+	(($#)) && echo "**** ERROR: $@"
 	# do not remove the entire WRKDIR in case sysmerge stopped half
 	# way since it contains our backup files
-	rm -f ${WRKDIR}/*SHA256*
+	rm -f ${WRKDIR}/{,x}etc-${SIGHASH}{,.sig}
 	rm -f ${WRKDIR}/*.tgz
 	rmdir ${WRKDIR} 2>/dev/null
 	exit 1
@@ -83,10 +89,14 @@ error_rm_wrkdir() {
 trap "restore_sum; clean_src; rm -rf ${WRKDIR}; exit 1" 1 2 3 13 15
 
 if (($(id -u) != 0)); then
-	error "need root privileges to run this script"
 	usage
-	error_rm_wrkdir
+	error_rm_wrkdir "need root privileges"
 fi
+
+# takes 2 arguments: output file and URL
+fetch() {
+	/usr/bin/ftp -V -m -k "${FTP_KEEPALIVE-0}" -o "$1" "$2" >/dev/null
+}
 
 # extract and verify (x)etcXX.tgz and create cksum file;
 # stores sum filename in ETCSUM or XETCSUM (see eval);
@@ -94,18 +104,12 @@ fi
 extract_set() {
 	[[ -z $1 ]] && return
 	local _tgz=$(readlink -f "$1") _set=$2 _f
-	local _sha256=${_set}-SHA256
-	if [ -z "${NOSIGCHECK}" ]; then
-		(cd ${WRKDIR} && sha256 -C ${_sha256} "${_tgz##*/}" >/dev/null 2>&1) || \
-			error_rm_wrkdir "${_tgz##*/} checksum could not be verified against ${_sha256}.sig"
-		rm "${WRKDIR}/${_sha256}"
-	fi
 	typeset -u _SETSUM=${_set}sum
 	eval ${_SETSUM}=${_set}sum
 	(cd ${TEMPROOT} && tar -xzphf "${_tgz}" && \
 		tar -tzf "${_tgz}" | while read _f; do
 			[ ! -h ${_f} ] && cksum ${_f} >> ${WRKDIR}/${_set}sum; done) || \
-				error_rm_wrkdir "extract/cksum of ${_tgz} failed"
+				error_rm_wrkdir "failed to extract ${_tgz} and create checksum file"
 	rm "${_tgz}"
 }
 
@@ -115,46 +119,55 @@ extract_set() {
 # takes url or filename and setname ('etc' or 'xetc') as arguments
 get_set() {
 	local _tgz=${WRKDIR}/${1##*/} _url=$1 _set=$2
-	local _sigfile=${_set}-SHA256
-	[ -f "${_url}" ] && _url="file://${_url}"
-	if [[ ${_url} == @(file|ftp|http|https)://*/*[!/] ]]; then 
-		${FETCH_CMD} -o ${_tgz} "${_url}" >/dev/null || \
-			error_rm_wrkdir "could not retrieve ${_url}"
+	local _sigfile=${WRKDIR}/${_set}-${SIGHASH}
+	[ -f "${_url}" ] && _url="file://$(readlink -f ${_url})"
+	if [[ ${_url} == @(file|ftp|http|https)://*/*[!/] ]]; then
+		echo "===> Fetching ${_url}"
+		fetch "${_tgz}" "${_url}" || \
+			error_rm_wrkdir "could not retrieve ${_url##*/}"
+	else
+			error_rm_wrkdir "${_url}: no such file"
 	fi
 	[[ ${_set} == etc ]] && TGZ=${_tgz} || XTGZ=${_tgz}
-	tar -tzf "${_tgz}" ./var/db/sysmerge/${_set}sum >/dev/null 2>&1 || \
-		error_rm_wrkdir "${_tgz} is not a valid ${_set}XX.tgz set"
+	tar -tzf "${_tgz}" ./var/db/sysmerge/${_set}sum >/dev/null || \
+		error_rm_wrkdir "${_tgz##*/}: badly formed \"${_set}\" set, lacks ./var/db/sysmerge/${_set}sum"
 	if [ -z "${NOSIGCHECK}" ]; then
-		${FETCH_CMD} -o "${WRKDIR}/${_sigfile}.sig" "${_url%/*}/SHA256.sig" >/dev/null 2>&1 || \
-			error_rm_wrkdir "could not retrieve ${_url%/*}/SHA256.sig"
-		check_sig "${_sigfile}"
+		echo "===> Fetching ${_url%/*}/${SIGHASH}.sig"
+		fetch "${_sigfile}.sig" "${_url%/*}/${SIGHASH}.sig" || \
+			error_rm_wrkdir "could not retrieve ${SIGHASH}.sig"
+		check_sig "${_sigfile}" "${_tgz}"
 	fi
 }
 
-# verify SHA256.sig and write ${WRKDIR}/(x)etc-SHA256, abort on failure;
-# takes a file path as argument
+# verify ${SIGHASH}.sig and write ${WRKDIR}/(x)etc-${SIGHASH}, abort on failure;
+# takes the signature file and set as arguments
 check_sig() {
-	local _sigfile=${WRKDIR}/$1 _src
-	local _key="/etc/signify/$(uname -r | tr -d '.')base.pub"
-	echo "===> Verying ${_set} SHA256 signature"
-	/usr/bin/signify -V -e -p ${_key} -x "${_sigfile}.sig" -m "${_sigfile}" 2>/dev/null || \
-		error_rm_wrkdir "signature check for ${_sigfile}"
-	rm "${_sigfile}.sig"
+	local _sigfile=${1##*/} _tgz=${2##*/}
+	local _key="/etc/signify/${RELINT}base.pub"
+	echo "===> Verifying ${_tgz} signature and checksum"
+	(cd ${WRKDIR} && \
+		signify -V -e -p ${_key} -x "${_sigfile}.sig" -m ${_sigfile} >/dev/null) || \
+			error_rm_wrkdir "${_sigfile}.sig: signature check failed"
+	(cd ${WRKDIR} && \
+		sha256 -C "${_sigfile}" "${_tgz}" >/dev/null) || \
+			error_rm_wrkdir "${_tgz}: bad ${SIGHASH} checksum"
+	rm ${WRKDIR}/${_sigfile}{,.sig}
 }
 
 # prepare TEMPROOT content from a src dir and create cksum file 
 prepare_src() {
 	[[ -z ${SRCDIR} ]] && return
 	SRCSUM=srcsum
+	# 2>/dev/null: distribution-etc-root-var complains /var/tmp is world writable
 	(cd ${SRCDIR}/etc && \
 	 make DESTDIR=${TEMPROOT} distribution-etc-root-var >/dev/null 2>&1 && \
 	 cd ${TEMPROOT} && find . -type f -and ! -type l | xargs cksum > ${WRKDIR}/${SRCSUM}) || \
-		error_rm_wrkdir "prepare/cksum of ${SRCDIR} failed"
+		error_rm_wrkdir "failed to populate from ${SRCDIR} and create checksum file"
 }
 
 sm_populate() {
 	local cf i _array _d _r _D _R CF_DIFF CF_FILES CURSUM IGNORE_FILES
-	mkdir -p ${DESTDIR}/${DBDIR} || error_rm_wrkdir
+	mkdir -p ${DESTDIR}/${DBDIR} || error_rm_wrkdir "cannot create ${DESTDIR}/${DBDIR}"
 	echo "===> Populating temporary root under ${TEMPROOT}"
 	mkdir -p ${TEMPROOT}
 
@@ -167,6 +180,7 @@ sm_populate() {
 			# delete file in temproot if it has not changed since last release
 			# and is present in current installation
 			if [ -z "${DIFFMODE}" ]; then
+				# 2>/dev/null: if file got removed manually but is still in the sum file
 				_R=$(cd ${TEMPROOT} && \
 					cksum -c ${DESTDIR}/${DBDIR}/${i} 2>/dev/null | awk '/OK/ { print $2 }' | sed 's/[:]//')
 				for _r in ${_R}; do
@@ -182,6 +196,7 @@ sm_populate() {
 			# set auto-upgradable files
 			_D=$(diff -u ${WRKDIR}/${i} ${DESTDIR}/${DBDIR}/${i} | grep -E '^\+' | sed '1d' | awk '{print $3}')
 			for _d in ${_D}; do
+				# 2>/dev/null: if file got removed manually but is still in the sum file
 				CURSUM=$(cd ${DESTDIR:=/} && cksum ${_d} 2>/dev/null)
 				[ -n "$(grep "${CURSUM}" ${DESTDIR}/${DBDIR}/${i})" -a -z "$(grep "${CURSUM}" ${WRKDIR}/${i})" ] && \
 					_array="${_array} ${_d}"
@@ -225,7 +240,7 @@ install_and_rm() {
 		cp ${5}/${4##*/} ${BKPDIR}/${4%/*}
 	fi
 
-	if ! install -m "${1}" -o "${2}" -g "${3}" "${4}" "${5}" 2>/dev/null; then
+	if ! install -m "${1}" -o "${2}" -g "${3}" "${4}" "${5}"; then
 		rm -f ${BKPDIR}/${4%/*}/${4##*/}
 		return 1
 	fi
@@ -320,11 +335,6 @@ merge_loop() {
 			case "${INSTALL_MERGED}" in
 			[eE])
 				echo "editing merged file...\n"
-				if [ -z "${VISUAL}" ]; then
-					EDIT="${EDITOR:=/usr/bin/vi}"
-				else
-					EDIT="${VISUAL}"
-				fi
 				${EDIT} ${COMPFILE}.merged
 				INSTALL_MERGED=v
 				;;
@@ -560,7 +570,7 @@ sm_compare() {
 	local _c1 _c2 _c3 COMPFILE CVSID1 CVSID2
 	echo "===> Starting comparison"
 
-	cd ${TEMPROOT} || error_rm_wrkdir
+	cd ${TEMPROOT} || error_rm_wrkdir "cannot enter ${TEMPROOT}"
 
 	# group and master.passwd need to be handled first in case
 	# install_file needs a new user/group;
@@ -574,7 +584,8 @@ sm_compare() {
 		# only process them (i.e. install) if they don't exist on the target system
 		if [ ! -s "${COMPFILE}" ]; then
 			if [ -f "${DESTDIR}${COMPFILE#.}" ]; then
-				rm "${COMPFILE}"
+				# group and master.passwd are always in the _c1 list
+				[ -f "${COMPFILE}" ] && rm "${COMPFILE}"
 			else
 				IS_BINFILE=1
 			fi
@@ -613,7 +624,7 @@ sm_compare() {
 
 		if [ -f "${COMPFILE}" -a -z "${IS_LINK}" ]; then
 			# make sure files are different; if not, delete the one in temproot
-			if diff -q "${DESTDIR}${COMPFILE#.}" "${COMPFILE}" >/dev/null 2>&1; then
+			if diff -q "${DESTDIR}${COMPFILE#.}" "${COMPFILE}" >/dev/null; then
 				rm "${COMPFILE}"
 			# xetcXX.tgz contains binary files; set IS_BINFILE to disable sdiff
 			elif diff -q "${DESTDIR}${COMPFILE#.}" "${COMPFILE}" | grep -q Binary; then
@@ -628,10 +639,9 @@ sm_compare() {
 
 sm_post() {
 	local FILES_IN_TEMPROOT FILES_IN_BKPDIR
-	echo "===> Checking directory hierarchy permissions (running mtree(8))"
-	mtree -qdef ${DESTDIR}/etc/mtree/4.4BSD.dist -p ${DESTDIR:=/} -U >/dev/null
-	[ -n "${XTGZ}" ] && \
-		mtree -qdef ${DESTDIR}/etc/mtree/BSD.x11.dist -p ${DESTDIR:=/} -U >/dev/null
+
+	FILES_IN_TEMPROOT=$(find ${TEMPROOT} -type f ! -name \*.merged -size +0)
+	[ -d "${BKPDIR}" ] && FILES_IN_BKPDIR=$(find ${BKPDIR} -type f -size +0)
 
 	if [ -n "${NEED_NEWALIASES}" ]; then
 		report "===> A new ${DESTDIR}/etc/mail/aliases file was installed."
@@ -639,8 +649,6 @@ sm_post() {
 		report "you will need to rebuild your aliases database manually.\n"
 	fi
 
-	FILES_IN_TEMPROOT=$(find ${TEMPROOT} -type f ! -name \*.merged -size +0 2>/dev/null)
-	FILES_IN_BKPDIR=$(find ${BKPDIR} -type f -size +0 2>/dev/null)
 	if [ -n "${AUTO_INSTALLED_FILES}" ]; then
 		report "===> Automatically installed file(s)"
 		report "${AUTO_INSTALLED_FILES}"
@@ -664,15 +672,6 @@ sm_post() {
 		report "${FILES_IN_TEMPROOT}"
 	fi
 
-	if [ -e "${REPORT}" ]; then
-		echo "===> Output log available at ${REPORT}"
-		find ${TEMPROOT} -type f -empty 2>/dev/null | xargs -r rm
-		find ${TEMPROOT} -type d | sort -r | xargs -r rmdir 2>/dev/null
-	else
-		echo "===> Removing ${WRKDIR}"
-		rm -rf "${WRKDIR}"
-	fi
-
 	[ -n "${FILES_IN_TEMPROOT}" ] && \
 		warn "some files are still left for comparison"
 
@@ -681,6 +680,20 @@ sm_post() {
 
 	[ -n "${NEED_REBOOT}" ] && \
 		warn "some new/updated file(s) may require a reboot"
+
+	echo "===> Checking directory hierarchy permissions (running mtree(8))"
+	mtree -qdef ${DESTDIR}/etc/mtree/4.4BSD.dist -p ${DESTDIR:=/} -U >/dev/null
+	[ -n "${XTGZ}" ] && \
+		mtree -qdef ${DESTDIR}/etc/mtree/BSD.x11.dist -p ${DESTDIR:=/} -U >/dev/null
+
+	if [ -e "${REPORT}" ]; then
+		echo "===> Output log available at ${REPORT}"
+		find ${TEMPROOT} -type f -empty | xargs -r rm
+		find ${TEMPROOT} -type d | sort -r | xargs -r rmdir 2>/dev/null
+	else
+		echo "===> Removing ${WRKDIR}"
+		rm -rf "${WRKDIR}"
+	fi
 
 	unset NEED_NEWALIASES NEED_REBOOT
 
@@ -701,7 +714,7 @@ while getopts bds:Sx: arg; do
 		if [ -d "${OPTARG}" ]; then
 			SRCDIR=${OPTARG}
 			[ -f "${SRCDIR}/etc/Makefile" ] || \
-				error_rm_wrkdir "${SRCDIR} is not a valid path to src"
+				error_rm_wrkdir "${SRCDIR}: invalid \"src\" tree, missing ${SRCDIR}/etc/Makefile"
 			continue
 		fi
 		get_set "${OPTARG}" etc
@@ -727,17 +740,15 @@ fi
 
 if [ -z "${SRCDIR}" -a -z "${TGZ}" -a -z "${XTGZ}" ]; then
 	if [ -n "${SM_PATH}" ]; then
-		_relint=$(uname -r | tr -d '.')
-		get_set "${SM_PATH}/etc${_relint}.tgz" etc
+		get_set "${SM_PATH}/etc${RELINT}.tgz" etc
 		if [ -d ${DESTDIR}/etc/X11 ]; then
-			get_set "${SM_PATH}/xetc${_relint}.tgz" xetc
+			get_set "${SM_PATH}/xetc${RELINT}.tgz" xetc
 		fi
 	elif [ -f "/usr/src/etc/Makefile" ]; then
 		SRCDIR=/usr/src
 	else
-		error "please specify a valid path to src or (x)etcXX.tgz"
 		usage
-		error_rm_wrkdir
+		error_rm_wrkdir "please specify a valid path to src or (x)etcXX.tgz"
 	fi
 fi
